requirement.txt
psycopg2
flask
openai
python-dotenv